Dummy implementation for discover account - adds support for JooQ as ORM - implements AccountDao for bulkInsert operation - all Response objects now implement BaseResponse - refactors Account fields based on PR discussion - adds correct enum for AccountType, AccountStatus, FiType, and ConsentType - moved implementation of all enums to source code instead of database
