--- src/main/java/org/indiastack/depa/provider/controllers/IndexController.java
@@ -1,15 +1,0 @@
-package org.indiastack.depa.provider.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/")
-public class IndexController {
-
-    @GetMapping
-    public String health() {
-        return "OK";
-    }
-}
