--- src/main/java/org/indiastack/depa/provider/dao/AccountDao.java
@@ -1,17 +1,0 @@
-package org.indiastack.depa.provider.dao;
-
-import org.indiastack.depa.provider.models.account.Account;
-import org.springframework.stereotype.Component;
-
-import java.util.ArrayList;
-
-@Component
-public class AccountDao {
-    public void save(Account account) {
-
-    }
-
-    public void bulkSave(ArrayList<Account> accounts) {
-
-    }
-}
