+++ src/main/java/org/indiastack/depa/provider/gateway/models/account/AccountStatus.java
@@ -1,0 +1,7 @@
+package org.indiastack.depa.provider.gateway.models.account;
+
+public enum AccountStatus {
+    DISCOVERED,
+    LINKED,
+    DE_LINKED
+}
