+++ src/main/java/org/indiastack/depa/provider/gateway/models/account/AccountType.java
@@ -1,0 +1,9 @@
+package org.indiastack.depa.provider.gateway.models.account;
+
+public enum AccountType {
+    SAVINGS,
+    CURRENT,
+    DEFAULT,
+    NRE,
+    NRO
+}
