+++ src/main/java/org/indiastack/depa/provider/gateway/models/account/http/AuthenticatorTypes.java
@@ -1,0 +1,5 @@
+package org.indiastack.depa.provider.gateway.models.account.http;
+
+public enum AuthenticatorTypes {
+    TOKEN
+}
