+++ src/main/java/org/indiastack/depa/provider/gateway/models/account/http/BaseAccountDetails.java
@@ -1,0 +1,31 @@
+package org.indiastack.depa.provider.gateway.models.account.http;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.indiastack.depa.provider.gateway.models.account.Account;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class BaseAccountDetails {
+    @JsonProperty("FIType")
+    private String fiType;
+
+    @JsonProperty("accType")
+    private String accountType;
+
+    @JsonProperty("accRefNumber")
+    private String accountRefNumber;
+
+    @JsonProperty("maskedAccNumber")
+    private String maskedAccountNumber;
+
+    public BaseAccountDetails(Account account) {
+        this.fiType = account.getType();
+
+        this.accountRefNumber = account.getBankRefNumber();
+        this.maskedAccountNumber = account.getNumber();
+    }
+}
