+++ src/main/java/org/indiastack/depa/provider/gateway/models/account/http/CustomerIdentifier.java
@@ -1,0 +1,14 @@
+package org.indiastack.depa.provider.gateway.models.account.http;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class CustomerIdentifier {
+    private String category;
+    private String type;
+    private String value;
+}
