+++ src/main/java/org/indiastack/depa/provider/gateway/models/commons/FiType.java
@@ -1,0 +1,23 @@
+package org.indiastack.depa.provider.gateway.models.commons;
+
+public enum FiType {
+    DEPOSIT,
+    TERM_DEPOSIT,
+    RECURRING_DEPOSIT,
+    SIP,
+    CP,
+    GOVT_SECURITIES,
+    EQUITIES,
+    BONDS,
+    DEBENTURES,
+    MUTUAL_FUNDS,
+    ETF,
+    IDR,
+    CIS,
+    AID,
+    INSURANCE_POLICIES,
+    NPS,
+    INVIT,
+    REIT,
+    OTHER
+}
