+++ src/main/java/org/indiastack/depa/provider/gateway/models/consent/ConsentFetchType.java
@@ -1,0 +1,6 @@
+package org.indiastack.depa.provider.gateway.models.consent;
+
+public enum ConsentFetchType {
+    ONE_TIME,
+    RECURRING
+}
