+++ src/main/java/org/indiastack/depa/provider/gateway/models/consent/ConsentStatus.java
@@ -1,0 +1,8 @@
+package org.indiastack.depa.provider.gateway.models.consent;
+
+public enum ConsentStatus {
+    INITIATED,
+    REVIEW,
+    GRANTED,
+    REVOKED
+}
