+++ src/main/java/org/indiastack/depa/provider/gateway/models/consent/ConsentType.java
@@ -1,0 +1,7 @@
+package org.indiastack.depa.provider.gateway.models.consent;
+
+public enum ConsentType {
+    PROFILE,
+    SUMMARY,
+    TRANSACTIONS
+}
