+++ src/main/java/org/indiastack/depa/provider/gateway/utils/DateUtils.java
@@ -1,0 +1,12 @@
+package org.indiastack.depa.provider.gateway.utils;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class DateUtils {
+    public static String getISODateInUTC() {
+        DateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
+        return format.format(new Date());
+    }
+}
