+++ src/main/java/org/indiastack/depa/provider/gateway/utils/DiscoverAccountUtils.java
@@ -1,0 +1,13 @@
+package org.indiastack.depa.provider.gateway.utils;
+
+import org.indiastack.depa.provider.integrations.banks.models.account.BankDiscoverRequest;
+import org.indiastack.depa.provider.gateway.models.account.http.DiscoverRequest;
+
+import java.util.ArrayList;
+
+public class DiscoverAccountUtils {
+
+    public static ArrayList<BankDiscoverRequest> createRequest(DiscoverRequest discoverRequest) {
+        return new ArrayList<>();
+    }
+}
