+++ src/main/java/org/indiastack/depa/provider/integrations/banks/models/account/BankDiscoverRequest.java
@@ -1,0 +1,4 @@
+package org.indiastack.depa.provider.integrations.banks.models.account;
+
+public class BankDiscoverRequest {
+}
