--- src/main/java/org/indiastack/depa/provider/models/account/AccountStatus.java
@@ -1,9 +1,0 @@
-package org.indiastack.depa.provider.models.account;
-
-public enum AccountStatus {
-    INITIATED,
-    IN_PROGRESS,
-    ACTIVE,
-    IN_ACTIVE,
-    CLOSED
-}
