--- src/main/java/org/indiastack/depa/provider/models/account/AccountType.java
@@ -1,7 +1,0 @@
-package org.indiastack.depa.provider.models.account;
-
-public enum AccountType {
-    SAVING,
-    CHECKING,
-    DEPOSIT
-}
