--- src/main/java/org/indiastack/depa/provider/models/account/http/AuthenticatorTypes.java
@@ -1,5 +1,0 @@
-package org.indiastack.depa.provider.models.account.http;
-
-public enum AuthenticatorTypes {
-    TOKEN
-}
