--- src/main/java/org/indiastack/depa/provider/models/account/http/BaseRequest.java
@@ -1,19 +1,0 @@
-package org.indiastack.depa.provider.models.account.http;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class BaseRequest {
-    @JsonProperty("ver")
-    private String version;
-
-    private String timestamp;
-
-    @JsonProperty("txnid")
-    private String transactionId;
-}
