--- src/main/java/org/indiastack/depa/provider/models/account/http/DiscoverRequest.java
@@ -1,25 +1,0 @@
-package org.indiastack.depa.provider.models.account.http;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class DiscoverRequest extends BaseRequest {
-    @JsonProperty("Customer")
-    private DiscoverCustomer discoverCustomer;
-
-    @JsonProperty("FITypes")
-    private ArrayList<String> fiTypes;
-
-    public DiscoverRequest(String version, String timestamp, String transactionId, DiscoverCustomer discoverCustomer, ArrayList<String> fiTypes) {
-        super(version, timestamp, transactionId);
-        this.discoverCustomer = discoverCustomer;
-        this.fiTypes = fiTypes;
-    }
-}
