--- src/main/java/org/indiastack/depa/provider/models/account/http/LinkRequest.java
@@ -1,18 +1,0 @@
-package org.indiastack.depa.provider.models.account.http;
-
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-public class LinkRequest extends BaseRequest {
-    @JsonProperty("Customer")
-    private LinkCustomer linkCustomer;
-
-    public LinkRequest(String version, String timestamp, String transactionId, LinkCustomer linkCustomer) {
-        super(version, timestamp, transactionId);
-        this.linkCustomer = linkCustomer;
-    }
-}
