--- src/main/java/org/indiastack/depa/provider/models/consent/ConsentFetchType.java
@@ -1,6 +1,0 @@
-package org.indiastack.depa.provider.models.consent;
-
-public enum ConsentFetchType {
-    ONE_TIME,
-    RECURRING
-}
