--- src/main/java/org/indiastack/depa/provider/models/consent/ConsentMode.java
@@ -1,6 +1,0 @@
-package org.indiastack.depa.provider.models.consent;
-
-public enum ConsentMode {
-    VIEW,
-    EDIT
-}
