--- src/main/java/org/indiastack/depa/provider/models/consent/ConsentStatus.java
@@ -1,8 +1,0 @@
-package org.indiastack.depa.provider.models.consent;
-
-public enum ConsentStatus {
-    INITIATED,
-    REVIEW,
-    GRANTED,
-    REVOKED
-}
