--- src/main/java/org/indiastack/depa/provider/services/ConsentService.java
@@ -1,12 +1,0 @@
-package org.indiastack.depa.provider.services;
-
-import org.indiastack.depa.provider.models.consent.Consent;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ConsentService {
-
-    public Consent createConsent(Consent consent) {
-        return new Consent();
-    }
-}
