--- src/main/java/org/indiastack/depa/provider/utils/DateUtils.java
@@ -1,12 +1,0 @@
-package org.indiastack.depa.provider.utils;
-
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-public class DateUtils {
-    public static String getISODateInUTC() {
-        DateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'");
-        return format.format(new Date());
-    }
-}
